Enable local copies of downloads

+++ pfb-analysis/import/import_osm.sh
@@ -58,11 +58,18 @@ BBOX_SW_LAT=`bc <<< "$BBOX_SW_LAT - $LAT_DIFF"`
 BBOX_SW_LNG=`bc <<< "$BBOX_SW_LNG - $LNG_DIFF"`
 BBOX_NE_LAT=`bc <<< "$BBOX_NE_LAT + $LAT_DIFF"`
 BBOX_NE_LNG=`bc <<< "$BBOX_NE_LNG + $LNG_DIFF"`
-# Download OSM data
-OSM_API_URL="http://www.overpass-api.de/api/xapi?*[bbox=${BBOX_SW_LNG},${BBOX_SW_LAT},${BBOX_NE_LNG},${BBOX_NE_LAT}]"
-OSM_TEMPDIR=`mktemp -d`
-OSM_DATA_FILE="${OSM_TEMPDIR}/overpass.osm"
-wget -O "${OSM_DATA_FILE}" "${OSM_API_URL}"
+
+if [[ -f ${1} ]]; then
+  echo "Importing provided OSM file"
+  OSM_DATA_FILE=${1}
+else
+  echo "Downloading OSM data"
+  # Download OSM data
+  OSM_API_URL="http://www.overpass-api.de/api/xapi?*[bbox=${BBOX_SW_LNG},${BBOX_SW_LAT},${BBOX_NE_LNG},${BBOX_NE_LAT}]"
+  OSM_TEMPDIR=`mktemp -d`
+  OSM_DATA_FILE="${OSM_TEMPDIR}/overpass.osm"
+  wget -O "${OSM_DATA_FILE}" "${OSM_API_URL}"
+fi
 
 # import the osm with highways
 osm2pgrouting \
